refactor(kepegawaian,portal): extract modular components for dashboard, cuti, and portal tabs (ponytail review)
