+++ graphify-out/cache/ast/v0.9.28/ebc3710920ad0873b9e6e3a629fc2805c05cb9d04b51e9ddb3996de0419ec935.json
@@ -1,0 +1,1 @@
+{"nodes": [{"id": "e_bksda_superapp_agents_agents_md", "label": "AGENTS.md", "file_type": "document", "source_file": ".agents/AGENTS.md", "source_location": "L1"}, {"id": "e_bksda_superapp_agents_agents_ponytail_lazy_senior_dev_mode", "label": "Ponytail, lazy senior dev mode", "file_type": "document", "source_file": ".agents/AGENTS.md", "source_location": "L1"}], "edges": [{"source": "e_bksda_superapp_agents_agents_md", "target": "e_bksda_superapp_agents_agents_ponytail_lazy_senior_dev_mode", "relation": "contains", "confidence": "EXTRACTED", "source_file": ".agents/AGENTS.md", "source_location": "L1", "weight": 1.0}], "input_tokens": 0, "output_tokens": 0}
